Rip out the collector mutex and rely on the GVL.
It's possible for the collector to yield the GVL whilst still holding
the collector lock, which leads to a deadlock. I _could_ fix this, but
actually since it turns out making ruby_memprofiler_pprof work with
Ractors is harder than I thought it was going to be anyway, rip out the
whole mutex infrastructure for now. We will probably need to do
something different to this when we support Ractors anyway.