Use the MJIT header + some copypasta, instead of debase
We actually need to reach in to the Ruby interpreter's rb_objspace
struct. The definition for this struct is not in any header file (it's
in gc.c), so we can't get it either from the MJIT header nor from
debase-ruby_core_source.

Instead, we have to copy-paste an appropriate version of that struct for
each Ruby version we support.

Since the layout of these structs depend on several compiler defines, we
need to know what Ruby was compiled with. Thankfully, a lot of the
salient bits (like the value of RGENGC_CHECK_MODE) are in the MJIT
header; so, bin debase-ruby_core_source, and use the MJIT header for
internal defs instead to pull those defines in.

We can get the value of some of the other defines by looking at GC::OPTS
during extconf time.