CI: Ensure CI runs on PRs from forks
In an open source project like this, contributors will typically create a fork
of the repository, create a new branch, push to that branch and open a pull
request on the "main" repository. In that workflow, the "push" event happens on
the forked repository while "pull_request" events happen on the main
repository. If we want CI results visible in the pull request, we need the jobs
to run in the main repository, hence we must run this workflow on the
"pull_request" event.

Since we still also want CI to run whenever something is pushed to the main
branch, e.g. when a pull request is merged, we configure CI to run on "push ->
branches -> main" too.